frost: split series processing into its own module
